+++ packages/manager/src/features/Billing/BillingPanels/BillingActivityPanel/BillingActivityPanel.tsx
@@ -14,7 +14,7 @@ import TableHead from 'src/components/core/TableHead';
 import Typography from 'src/components/core/Typography';
 import { TextTooltip } from 'src/components/TextTooltip';
 import { Currency } from 'src/components/Currency';
-import DateTimeDisplay from 'src/components/DateTimeDisplay';
+import { DateTimeDisplay } from 'src/components/DateTimeDisplay';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
 import InlineMenuAction from 'src/components/InlineMenuAction';
 import Link from 'src/components/Link';